+++ src/components/Main/Heading/MobilePetsHeading.tsx
