+++ src/components/Main/Heading/PetsHeading.tsx
